refactor: css 파일 위치 변경
atom 폴더에 css 파일이 같이 있는 것이 불편하다 판단되어 css 폴더를 따로 만들어 css 파일을 분리함.
